LF-4747 Move into SensorTile folder - tbd if moving up to components level
